requirements.txt
pyserial==3.4
python-uinput==0.11.2
mrcrowbar==0.4.1

